test: 렌주룰 위반 테스트 Player로 이동
